Added mixin for resume (used in job seeker signup and apply forms)
