+++ itou/www/apply/forms.py
@@ -1,12 +1,8 @@
 import datetime
-import urllib
-from urllib import request
 
-import requests
 from dateutil.relativedelta import relativedelta
 from django import forms
 from django.contrib.auth import get_user_model
-from django.core.validators import URLValidator
 from django.db.models import Q
 from django.utils import timezone
 from django.utils.translation import gettext as _, gettext_lazy
@@ -16,6 +12,7 @@ from itou.approvals.models import Approval
 from itou.job_applications.models import JobApplication, JobApplicationWorkflow
 from itou.prescribers.models import PrescriberOrganization
 from itou.utils.address.forms import AddressFormMixin
+from itou.utils.resume.forms import ResumeFormMixin
 from itou.utils.widgets import DatePickerField
 
 
@@ -64,7 +61,7 @@ class CheckJobSeekerInfoForm(forms.ModelForm):
         )
 
 
-class CreateJobSeekerForm(AddressFormMixin, forms.ModelForm):
+class CreateJobSeekerForm(AddressFormMixin, ResumeFormMixin, forms.ModelForm):
     def __init__(self, proxy_user, *args, **kwargs):
         self.proxy_user = proxy_user
         super().__init__(*args, **kwargs)
@@ -77,10 +74,6 @@ class CreateJobSeekerForm(AddressFormMixin, forms.ModelForm):
         self.fields["birthdate"].required = True
         self.fields["birthdate"].widget = DatePickerField()
         self.fields["birthdate"].input_formats = [DatePickerField.DATE_FORMAT]
-
-        # Resume URL
-        self.fields["resume_link"].widget.attrs["placeholder"] = gettext_lazy("Entrez l'adresse de votre CV")
-        self.fields["resume_link"].required = False
 
     class Meta:
         model = get_user_model()
@@ -109,23 +102,6 @@ class CreateJobSeekerForm(AddressFormMixin, forms.ModelForm):
         if get_user_model().email_already_exists(email):
             raise forms.ValidationError(get_user_model().ERROR_EMAIL_ALREADY_EXISTS)
         return email
-
-    def clean_resume_link(self):
-        resume_link = self.cleaned_data["resume_link"]
-
-        if resume_link:
-            # Check if valid (also on model)
-            validator = URLValidator(schemes=["https"])
-            validator(resume_link)
-
-            # Check if exists ?
-            try:
-                requests.head(resume_link, allow_redirects=True)
-            except requests.exceptions.RequestException:
-                # Catch all
-                raise forms.ValidationError(gettext_lazy("Cette URL n'est pas accessible"))
-
-        return resume_link
 
     def clean(self):
         super().clean()
